+++ Tests/CollectionTests.swift
@@ -31,9 +31,9 @@ class CollectionTests: XCTestCase {
     
     func testCount() {
         
-        XCTAssertEqual([1, 2, 0, -1, 3].count({ $0 > 0 }), 3)
-        XCTAssertEqual([0, 1, 2, 0, -1].count({ $0 > 0 }), 2)
-        XCTAssertEqual([1, 2, 3, 4, 5].count({ $0 > 0 }), 5)
+        XCTAssertEqual([1, 2, 0, -1, 3].count(where: { $0 > 0 }), 3)
+        XCTAssertEqual([0, 1, 2, 0, -1].count(where: { $0 > 0 }), 2)
+        XCTAssertEqual([1, 2, 3, 4, 5].count(where: { $0 > 0 }), 5)
         
         XCTAssertEqual([1, 2, 0, -1, 3].countPrefix(while: { $0 > 0 }), 2)
         XCTAssertEqual([0, 1, 2, 0, -1].countPrefix(while: { $0 > 0 }), 0)